pyproject.toml
[project]
name = "faiss-rag-system"
version = "0.1.0"
description = "FAISS RAG system with Redis, Ollama, FastAPI, and hot-reloadable plugins"
authors = [{ name = "SGloop", email = "dev@custard.solutions" }]
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.110.0",
    "pydantic>=2.0",
    "httpx",
    "redis>=5.0.0",
    "uvicorn[standard]",
    "jinja2",
    "faiss-cpu"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "redis"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
